3rd commit, after merge with PO branche
volledig gemerged met PO branche, werkt als een trein

+++ BataviaReseveringsSysteem/Views/BoatList.xaml.cs
@@ -58,6 +58,12 @@ namespace Views
         void ButtonDelete(object sender, RoutedEventArgs e)
         {
             Button b = (Button)sender;
+            if (MessageBox.Show("Wilt u deze boot definitief verwijderen?",
+                    "Bevestig verwijdering",
+                    MessageBoxButton.YesNo,
+                    MessageBoxImage.Question) !=
+                MessageBoxResult.Yes)
+                return;
             bc.DeleteBoat((int)b.Tag);
             Switcher.Switch(new BoatList());
         }
